Add tcmalloc
This was born out of a discussion between us discussing the removal of
alloc_aligned() and free_aligned() from the HSE codebase. In that
discussion, we looked a bit at alternative allocators. tcmalloc proved
to be the most performant allocator based on Greg's little bit of
testing.

tcmalloc will (hopefully) become an optional dependency of HSE, so we
can drop our custom aligned allcator API.

Signed-off-by: Tristan Partin <tpartin@micron.com>